requirements.txt
numpy==1.20.3
keras=2.7.0
opencv-python==4.5.4.60
tensorflow==2.7.0
pillow==8.4.0
matplotlib==3.4.3
